+++ packages/falcon-client/build-utils/webpack/config/env.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const paths = require('./../../paths');
 
 function getClientEnv(target, options, envToBuildIn) {
@@ -16,9 +17,9 @@ function getClientEnv(target, options, envToBuildIn) {
         ...(env === 'development' ? { DEV_SERVER_PORT: devServerPort } : {}),
         BUILD_TARGET: target === 'web' ? 'client' : 'server',
         PUBLIC_PATH: env === 'production' ? publicPath : undefined,
-        ASSETS_MANIFEST: paths.appManifest,
-        LOADABLE_STATS: paths.appLoadableStats,
-        PUBLIC_DIR: env === 'production' ? paths.appBuildPublic : paths.appPublic
+        WEBPACK_ASSETS: paths.appWebpackAssets,
+        OUTPUT_DIR: path.relative(paths.appPath, paths.appBuildPublic),
+        PUBLIC_DIR: path.relative(paths.appPath, env === 'production' ? paths.appBuildPublic : paths.appPublic)
       }
     );