+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 - `menu`, `Category`, `Category.products`, `Product` and `Product.breadcrumbs` resolvers are cached now ([#374](https://github.com/deity-io/falcon/pull/374))
 
+### Falcon Client (Next)
+
+- fixed absolute path to output directory embedded in bundle ([385](https://github.com/deity-io/falcon/issues/385))
+
 ### Falcon E-commerce UI Kit (Next)
 
 - added basic component set (queries, mutations and ui) to build filters ([#365](https://github.com/deity-io/falcon/pull/365))